+++ R/archive.R
@@ -114,7 +114,7 @@ writeArchive <- function(object, file) {
   is.character(file) || stop("in archive(object, file): 'file' must be supplied as a string. ",
                              "\nDid you omit the surrounding quotes \"\"?", call. = FALSE)
 
-  if (file.exists(file)) {
+  if (file.exists(file) && interactive()) {
     continue <- readline(prompt = paste0("The file '", file, "' already exists. Overwrite? [y/N] "))
     if (!continue %in% c("y", "Y", "yes", "YES")) {
       message("Request cancelled")